// schemas/public/BlogPostCategory.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { z } from 'zod';

/** Identifier type for public.blog_post_category */
export type BlogPostCategoryId = string & { __brand: 'BlogPostCategoryId' };

/** Represents the table public.blog_post_category */
export default interface BlogPostCategory {
  id: BlogPostCategoryId;

  identifier: string;

  label: string;

  description: string | null;

  parent_id: BlogPostCategoryId | null;

  seo_title: string;

  seo_description: string | null;

  seo_keywords: string | null;

  x_created: Date | null;
}

/** Represents the initializer for the table public.blog_post_category */
export interface BlogPostCategoryInitializer {
  /** Default value: uuid_generate_v4() */
  id?: BlogPostCategoryId;

  identifier: string;

  label: string;

  description?: string | null;

  parent_id?: BlogPostCategoryId | null;

  seo_title: string;

  seo_description?: string | null;

  seo_keywords?: string | null;

  /** Default value: CURRENT_TIMESTAMP */
  x_created?: Date | null;
}

/** Represents the mutator for the table public.blog_post_category */
export interface BlogPostCategoryMutator {
  id?: BlogPostCategoryId;

  identifier?: string;

  label?: string;

  description?: string | null;

  parent_id?: BlogPostCategoryId | null;

  seo_title?: string;

  seo_description?: string | null;

  seo_keywords?: string | null;

  x_created?: Date | null;
}

export const blogPostCategoryId =
  z.string() as unknown as z.Schema<BlogPostCategoryId>;

export const blogPostCategory = z.object({
  id: blogPostCategoryId,
  identifier: z.string(),
  label: z.string(),
  description: z.string().nullable(),
  parent_id: blogPostCategoryId.nullable(),
  seo_title: z.string(),
  seo_description: z.string().nullable(),
  seo_keywords: z.string().nullable(),
  x_created: z.date().nullable(),
}) as unknown as z.Schema<BlogPostCategory>;

export const blogPostCategoryInitializer = z.object({
  id: blogPostCategoryId.optional(),
  identifier: z.string(),
  label: z.string(),
  description: z.string().optional().nullable(),
  parent_id: blogPostCategoryId.optional().nullable(),
  seo_title: z.string(),
  seo_description: z.string().optional().nullable(),
  seo_keywords: z.string().optional().nullable(),
  x_created: z.date().optional().nullable(),
}) as unknown as z.Schema<BlogPostCategoryInitializer>;

export const blogPostCategoryMutator = z.object({
  id: blogPostCategoryId.optional(),
  identifier: z.string().optional(),
  label: z.string().optional(),
  description: z.string().optional().nullable(),
  parent_id: blogPostCategoryId.optional().nullable(),
  seo_title: z.string().optional(),
  seo_description: z.string().optional().nullable(),
  seo_keywords: z.string().optional().nullable(),
  x_created: z.date().optional().nullable(),
}) as unknown as z.Schema<BlogPostCategoryMutator>;